refactor: Extract function to get type of entries
