+++ src/drive/locales/en.json
@@ -580,11 +580,6 @@
     "cancelledWithRestoreErrors": "%{subject} has been moved back to it's original location but there was an error while restoring the file from trash. |||| %{smart_count} elements have been moved back to their original location but there was %{restoreErrorsCount} error(s) while restoring the file(s) from trash.",
     "cancelled_error": "Sorry, there was an error while moving the element back. |||| Sorry, there was an error while moving these elements back.",
     "multipleEntries": "%{smart_count} element |||| %{smart_count} elements",
-    "entriesType": {
-      "file": "file",
-      "directory": "folder",
-      "element": "element"
-    },
     "outsideSharedFolder": {
       "title": "Moving outside the %{sharedFolder} folder",
       "content_1": "Warning, you want to move %{name} out of the shared %{sharedFolder} folder. |||| Warning, you want to move %{smart_count} %{type}s out of the shared %{sharedFolder} folder.",
@@ -767,5 +762,10 @@
     "restoreSuccess": "The item has been restored",
     "restoreError": "An error has occurred, please try again.",
     "destroySuccess": "The item has been deleted"
+  },
+  "EntriesType": {
+    "file": "file |||| files",
+    "directory": "folder |||| folders",
+    "element": "element |||| elements"
   }
 }
